feat: landing page, error handling all page, input validation

+++ delete.php
@@ -1,5 +1,12 @@
 <?php
 include "koneksi.php";
+
+session_start();
+if (!isset($_SESSION['id'])) {
+    header("location:login.php");
+    die;
+}
+
 if (isset($_GET) && $_GET['id']) {
     $id = $_GET['id'];
     $query = $conn->query("DELETE FROM obat WHERE id = $id");
